# Web APi/RavCevelGood1/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RavCevelGood",
    "version": "v1"
  },
  "paths": {
    "/api/Course/GetAllCourseBySchoolDAndYearbookId/{SchoolId}/{YearbookId}": {
      "get": {
        "tags": [
          "Course"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "YearbookId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Course/GetAllCourseBySchoolId/{SchoolId}": {
      "get": {
        "tags": [
          "Course"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Course/GetAllSemester/{YearbookId}": {
      "get": {
        "tags": [
          "Course"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "YearbookId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Course/DeleteCourse/{GroupSemesterPerCourseId}": {
      "delete": {
        "tags": [
          "Course"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GroupSemesterPerCourseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Course/GetUsersPerCourse/{CourseId}": {
      "get": {
        "tags": [
          "Course"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "CourseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Group/GetGroupsByIdSchool/{SchoolId}/{YearbookId}": {
      "get": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "YearbookId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Group/GetStudentPerGroup/{GroupId}/{YearbookId}": {
      "get": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GroupId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "YearbookId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Group/DeleteGroup/{GroupId}/{YearbookId}": {
      "delete": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GroupId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "YearbookId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Group/AddGroup/{UserCreatedId}/{SchoolId}/{YearbookId}/{userId}": {
      "put": {
        "tags": [
          "Group"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "UserCreatedId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "YearbookId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AppGroupDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Group/GetAllNameGroup/{SchoolId}": {
      "get": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AppUserPerGroupDTO": {
      "type": "object",
      "properties": {
        "iduserPerGroup": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "groupId": {
          "format": "int32",
          "type": "integer"
        },
        "fromDate": {
          "format": "date-time",
          "type": "string"
        },
        "toDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "AppGroupPerYearbookDTO": {
      "type": "object",
      "properties": {
        "idgroupPerYearbook": {
          "format": "int32",
          "type": "integer"
        },
        "groupId": {
          "format": "int32",
          "type": "integer"
        },
        "yearbookId": {
          "format": "int32",
          "type": "integer"
        },
        "appUserPerGroups": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AppUserPerGroupDTO"
          }
        }
      }
    },
    "AppGroupDTO": {
      "type": "object",
      "properties": {
        "idgroup": {
          "format": "int32",
          "type": "integer"
        },
        "ageGroupId": {
          "format": "int32",
          "type": "integer"
        },
        "schoolId": {
          "format": "int32",
          "type": "integer"
        },
        "typeGroupId": {
          "format": "int32",
          "type": "integer"
        },
        "extensionId": {
          "format": "int32",
          "type": "integer"
        },
        "voiceSpaceId": {
          "format": "int32",
          "type": "integer"
        },
        "listeningTimeId": {
          "format": "int32",
          "type": "integer"
        },
        "userCreated": {
          "format": "int32",
          "type": "integer"
        },
        "dateCreated": {
          "format": "date-time",
          "type": "string"
        },
        "userUpdated": {
          "format": "int32",
          "type": "integer"
        },
        "dateUpdated": {
          "format": "date-time",
          "type": "string"
        },
        "nameGroup": {
          "type": "string"
        },
        "idgroupPeryearbook": {
          "format": "int32",
          "type": "integer"
        },
        "appGroupPerYearbooks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AppGroupPerYearbookDTO"
          }
        }
      }
    },
...
  }
}